+++ like/src/events.ts
@@ -13,6 +13,9 @@ declare global {
     ['like:mousemoved']: CustomEvent<{pos: Vector2, delta: Vector2}>;
     ['like:updateRenderTarget']: CustomEvent<{target: HTMLCanvasElement}>;
     ['like:resizeCanvas']: CustomEvent<{size: Vector2}>;
+    ['like:preDraw']: CustomEvent<{}>;
+    ['like:postDraw']: CustomEvent<{}>;
+    ['like:update']: CustomEvent<{dt: number}>;
   }
 }